requirements.txt
torchinfo
protobuf==3.20
exceptiongroup
datasets
wandb
fire
scipy
pandas
scikit-learn
tabulate